FontREPLEACE

+++ src/minecraft/me/skrilled/utils/render/gl/GLColor.java
@@ -14,6 +14,6 @@ public class GLColor {
 		float red = (float) (color >> 16 & 255) / 255.0f;
 		float green = (float) (color >> 8 & 255) / 255.0f;
 		float blue = (float) (color & 255) / 255.0f;
-		GL11.glColor4f((float) red, (float) green, (float) blue, (float) alpha);
+		GL11.glColor4f(red, green, blue, alpha);
 	}
 }
