+++ src/minecraft/net/minecraft/client/renderer/block/model/BreakingFour.java
@@ -1,7 +1,8 @@
 package net.minecraft.client.renderer.block.model;
 
-import java.util.Arrays;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
+
+import java.util.Arrays;
 
 public class BreakingFour extends BakedQuad
 {
@@ -65,7 +66,7 @@ public class BreakingFour extends BakedQuad
                 f4 = (1.0F - f1) * 16.0F;
         }
 
-        this.vertexData[j + 4] = Float.floatToRawIntBits(this.texture.getInterpolatedU((double)f3));
-        this.vertexData[j + 4 + 1] = Float.floatToRawIntBits(this.texture.getInterpolatedV((double)f4));
+        this.vertexData[j + 4] = Float.floatToRawIntBits(this.texture.getInterpolatedU(f3));
+        this.vertexData[j + 4 + 1] = Float.floatToRawIntBits(this.texture.getInterpolatedV(f4));
     }
 }
