+++ src/minecraft/net/minecraft/client/renderer/entity/RenderLightningBolt.java
@@ -1,12 +1,13 @@
 package net.minecraft.client.renderer.entity;
 
-import java.util.Random;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.renderer.WorldRenderer;
 import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
 import net.minecraft.entity.effect.EntityLightningBolt;
 import net.minecraft.util.ResourceLocation;
+
+import java.util.Random;
 
 public class RenderLightningBolt extends Render<EntityLightningBolt>
 {
@@ -33,8 +34,8 @@ public class RenderLightningBolt extends Render<EntityLightningBolt>
         {
             adouble[i] = d0;
             adouble1[i] = d1;
-            d0 += (double)(random.nextInt(11) - 5);
-            d1 += (double)(random.nextInt(11) - 5);
+            d0 += random.nextInt(11) - 5;
+            d1 += random.nextInt(11) - 5;
         }
 
         for (int k1 = 0; k1 < 4; ++k1)
@@ -64,15 +65,13 @@ public class RenderLightningBolt extends Render<EntityLightningBolt>
                     double d4 = d2;
                     double d5 = d3;
 
-                    if (j == 0)
-                    {
-                        d2 += (double)(random1.nextInt(11) - 5);
-                        d3 += (double)(random1.nextInt(11) - 5);
+                    if (j == 0) {
+                        d2 += random1.nextInt(11) - 5;
+                        d3 += random1.nextInt(11) - 5;
                     }
-                    else
-                    {
-                        d2 += (double)(random1.nextInt(31) - 15);
-                        d3 += (double)(random1.nextInt(31) - 15);
+                    else {
+                        d2 += random1.nextInt(31) - 15;
+                        d3 += random1.nextInt(31) - 15;
                     }
 
                     worldrenderer.begin(5, DefaultVertexFormats.POSITION_COLOR);
