+++ src/minecraft/net/minecraft/client/renderer/tileentity/TileEntityPistonRenderer.java
@@ -6,11 +6,7 @@ import net.minecraft.block.BlockPistonExtension;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.BlockRendererDispatcher;
-import net.minecraft.client.renderer.GlStateManager;
-import net.minecraft.client.renderer.RenderHelper;
-import net.minecraft.client.renderer.Tessellator;
-import net.minecraft.client.renderer.WorldRenderer;
+import net.minecraft.client.renderer.*;
 import net.minecraft.client.renderer.texture.TextureMap;
 import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
 import net.minecraft.init.Blocks;
@@ -48,7 +44,7 @@ public class TileEntityPistonRenderer extends TileEntitySpecialRenderer<TileEnti
             }
 
             worldrenderer.begin(7, DefaultVertexFormats.BLOCK);
-            worldrenderer.setTranslation((double)((float)x - (float)blockpos.getX() + te.getOffsetX(partialTicks)), (double)((float)y - (float)blockpos.getY() + te.getOffsetY(partialTicks)), (double)((float)z - (float)blockpos.getZ() + te.getOffsetZ(partialTicks)));
+            worldrenderer.setTranslation((float) x - (float) blockpos.getX() + te.getOffsetX(partialTicks), (float) y - (float) blockpos.getY() + te.getOffsetY(partialTicks), (float) z - (float) blockpos.getZ() + te.getOffsetZ(partialTicks));
             World world = this.getWorld();
 
             if (block == Blocks.piston_head && te.getProgress(partialTicks) < 0.5F)
@@ -62,7 +58,7 @@ public class TileEntityPistonRenderer extends TileEntitySpecialRenderer<TileEnti
                 IBlockState iblockstate1 = Blocks.piston_head.getDefaultState().withProperty(BlockPistonExtension.TYPE, blockpistonextension$enumpistontype).withProperty(BlockPistonExtension.FACING, iblockstate.getValue(BlockPistonBase.FACING));
                 iblockstate1 = iblockstate1.withProperty(BlockPistonExtension.SHORT, Boolean.valueOf(te.getProgress(partialTicks) >= 0.5F));
                 this.blockRenderer.getBlockModelRenderer().renderModel(world, this.blockRenderer.getModelFromBlockState(iblockstate1, world, blockpos), iblockstate1, blockpos, worldrenderer, true);
-                worldrenderer.setTranslation((double)((float)x - (float)blockpos.getX()), (double)((float)y - (float)blockpos.getY()), (double)((float)z - (float)blockpos.getZ()));
+                worldrenderer.setTranslation((float) x - (float) blockpos.getX(), (float) y - (float) blockpos.getY(), (float) z - (float) blockpos.getZ());
                 iblockstate.withProperty(BlockPistonBase.EXTENDED, Boolean.valueOf(true));
                 this.blockRenderer.getBlockModelRenderer().renderModel(world, this.blockRenderer.getModelFromBlockState(iblockstate, world, blockpos), iblockstate, blockpos, worldrenderer, true);
             }
