+++ src/minecraft/net/minecraft/nbt/NBTTagInt.java
@@ -63,7 +63,7 @@ public class NBTTagInt extends NBTBase.NBTPrimitive
 
     public long getLong()
     {
-        return (long)this.data;
+        return this.data;
     }
 
     public int getInt()
@@ -83,7 +83,7 @@ public class NBTTagInt extends NBTBase.NBTPrimitive
 
     public double getDouble()
     {
-        return (double)this.data;
+        return this.data;
     }
 
     public float getFloat()
